+++ com.github.sormuras.bach/main/java/com/github/sormuras/bach/project/CodeStyle.java
@@ -1,7 +1,7 @@
 package com.github.sormuras.bach.project;
 
 /** Specify the style to format/verify Java sources. */
-public enum JavaStyle {
+public enum CodeStyle {
   /** Be free, have fun - use spaces and tabs. */
   FREE,
   /** <a href="https://google.github.io/styleguide/javaguide.html">Google Java Style</a>. */