+++ com.github.sormuras.bach/main/java/com/github/sormuras/bach/project/Spaces.java
@@ -1,4 +1,4 @@
 package com.github.sormuras.bach.project;
 
 /** A record of code spaces and code related properties. */
-public record Spaces(JavaStyle style, MainSpace main, TestSpace test) {}
+public record Spaces(CodeStyle style, MainSpace main, TestSpace test) {}