+++ CHANGELOG.md
@@ -14,6 +14,7 @@ _In progress..._
 
 - Generator methods in class `Command` are now called `with()` and `withAll()` instead of `.add()` `addAll()`.
   This also affects all implementations, like `Javac`, `Javadoc`. [#205]
+- `CodeStyle` renamed from `JavaStyle`. [#205]
 
 ### Fixed